requirements.txt
discord.py
python-dotenv
gdown
torch
torchvision
pandas
pillow